fix(TUP-31736): fix warning message
